Invalid Date Fix
Mismatch query on date field
